Extract base ogds listing service.
