+++ opengever/api/userlisting.py
@@ -1,20 +1,9 @@
-from opengever.base.model import create_session
-from opengever.base.utils import safe_int
+from opengever.api.ogdslistingbase import OgdsListingBaseService
 from opengever.ogds.models.user import User
-from opengever.tabbedview.sqlsource import cast_to_string
-from opengever.tabbedview.sqlsource import sort_column_exists
-from plone.restapi.batching import HypermediaBatch
-from plone.restapi.services import Service
-from Products.CMFPlone.utils import safe_unicode
-from sqlalchemy import or_
-from sqlalchemy.sql.expression import asc
-from sqlalchemy.sql.expression import column
-from sqlalchemy.sql.expression import desc
-from zExceptions import BadRequest
 from ZPublisher.HTTPRequest import record
 
 
-class UserListingGet(Service):
+class UserListingGet(OgdsListingBaseService):
     """API Endpoint that returns users from ogds.
 
     GET /@user-listing HTTP/1.1
@@ -46,88 +35,19 @@ class UserListingGet(Service):
     )
 
     default_sort_on = 'lastname'
-    default_sort_order = 'ascending'
+    model_class = User
     default_state_filter = tuple()
 
-    def reply(self):
-        params = self.request.form.copy()
-
-        session = create_session()
-        query = session.query(User)
-        query = self.extend_query_with_sorting(query, params)
-        query = self.extend_query_with_search(query, params)
-        query = self.extend_query_with_filters(query, params)
-        items_total = query.count()
-        b_start, b_size, query = self.extend_query_with_batching(query, params)
-
-        items = []
-        for user in query.all():
-            item = {}
-            for colname in self.item_columns:
-                item[colname] = getattr(user, colname)
-            item['@type'] = 'virtual.ogds.user'
-            item['title'] = user.fullname()
-            # currently a dummy link but required by the UI
-            item['@id'] = '{}/@ogds-user/{}'.format(
-                self.context.absolute_url(), user.userid)
-            items.append(item)
-
-        # We use HypermediaBatch for the canonical url only
-        batch = HypermediaBatch(self.request, items)
-        # return empty facet dict to keep response structure consistent
-        return {
-          "@id": batch.canonical_url,
-          "b_size": b_size,
-          "b_start": b_start,
-          "facets": {},
-          "items": items,
-          "items_total": items_total
-        }
-
-    def extend_query_with_sorting(self, query, params):
-        sort_on = params.get('sort_on', self.default_sort_on).strip()
-        sort_order = params.get('sort_order', self.default_sort_order)
-
-        # early abort if the column is not in the query
-        if not sort_column_exists(query, sort_on):
-            return query
-
-        # Don't plug column names as literal strings into an order_by
-        # clause, but use a ColumnClause instead to allow SQLAlchemy to
-        # properly quote the identifier name depending on the dialect
-        sort_on = column(sort_on)
-
-        if sort_order in ['descending', 'reverse']:
-            order_f = desc
-        else:
-            order_f = asc
-        return query.order_by(order_f(sort_on))
-
-    def extend_query_with_search(self, query, params):
-        search = params.get('search', '').strip()
-        if not search:
-            return query
-
-        search = safe_unicode(search)
-
-        # remove trailing asterisk
-        if search.endswith(u'*'):
-            search = search[:-1]
-
-        # split up the search term into words, extend them with the default
-        # wildcards and then search for every word seperately
-        for word in search.split():
-            term = u'%%%s%%' % word
-
-            expressions = []
-            for field in self.searchable_columns:
-                expressions.append(cast_to_string(field).ilike(term))
-            query = query.filter(or_(*expressions))
-
-        return query
+    def fill_item(self, item, model):
+        item = super(UserListingGet, self).fill_item(item, model)
+        item['@type'] = 'virtual.ogds.user'
+        item['title'] = model.fullname()
+        item['@id'] = '{}/@ogds-user/{}'.format(
+            self.context.absolute_url(), model.userid)
+        return item
 
     def extend_query_with_filters(self, query, params):
-        """Handle hardcoded state filter.
+        """Implement hardcoded state filter.
 
         The state filter expects a list of states to be displayed. By default
         it will return all states (active and inacvite users).
@@ -142,15 +62,3 @@ class UserListingGet(Service):
         elif state == ['inactive']:
             query = query.filter_by(active=False)
         return query
-
-    def extend_query_with_batching(self, query, params):
-        b_start = safe_int(params.get('b_start', 0), 0)
-        if b_start < 0:
-            raise BadRequest("The parameter 'b_start' can't be negative.")
-        b_size = min(safe_int(params.get('b_size', 25), 25), 100)
-        if b_size < 0:
-            raise BadRequest("The parameter 'b_size' can't be negative.")
-
-        query = query.offset(b_start)
-        query = query.limit(b_size)
-        return b_start, b_size, query
